+++ shared-components/src/containers/Mainpage/containers/SingleTaskPage/WorkspaceView/Structures/TableView/TableCell/ColumnHeaderCell/ColumnHeaderCell.js
@@ -296,12 +296,14 @@ class ColumnHeaderCell extends Component {
   };
 
   deleteTableColumnByIndex = event => {
-    FirestoreManager.deleteColumnInTableByIndex(
-      this.props.workspace.id,
-      this.props.columnIndex
-    );
+    if (window.confirm(`Are you sure you want to delete this column?`)) {
+      FirestoreManager.deleteColumnInTableByIndex(
+        this.props.workspace.id,
+        this.props.columnIndex
+      );
 
-    this.props.setColumnToDelete(-1);
+      this.props.setColumnToDelete(-1);
+    }
   };
 
   addPieceToThisCell = pieceId => {
